chore: move analytics definitions to suite-common
