+++ packages/suite-analytics/tsconfig.json
@@ -1,5 +1,7 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": [{ "path": "../analytics" }]
+    "references": [
+        { "path": "../../suite-common/analytics" }
+    ]
 }
