+++ packages/suite/src/support/suite/useConnectPopupDesktop.tsx
@@ -7,7 +7,7 @@ import {
 } from '@suite-common/connect-popup';
 import { CALL_SOURCE_DESKTOP_WS } from '@suite-common/connect-popup/src/connectPopupTypes';
 import TrezorConnect from '@trezor/connect';
-import { EventType, analytics } from '@trezor/suite-analytics';
+import { EventTypeShared, analytics } from '@trezor/suite-analytics';
 import { desktopApi } from '@trezor/suite-desktop-api';
 
 import { SUITE } from 'src/actions/suite/constants';
@@ -56,7 +56,7 @@ export const useConnectPopupDesktop = () => {
                     desktopApi.connectPopupReady();
 
                     analytics.report({
-                        type: EventType.ConnectPopupInit,
+                        type: EventTypeShared.ConnectPopupInit,
                     });
 
                     // Update experimental features value
