+++ suite-common/analytics/package.json
@@ -13,6 +13,14 @@
     "dependencies": {
         "@reduxjs/toolkit": "2.8.2",
         "@suite-common/redux-utils": "workspace:*",
+        "@suite-common/suite-config": "workspace:*",
+        "@suite-common/suite-constants": "workspace:*",
+        "@suite-common/trading": "workspace:*",
+        "@suite-common/wallet-config": "workspace:*",
+        "@suite-common/wallet-types": "workspace:*",
+        "@trezor/analytics": "workspace:*",
+        "@trezor/connect": "workspace:*",
+        "@trezor/env-utils": "workspace:*",
         "react": "18.2.0"
     }
 }
