+++ suite-common/analytics/src/events/shared/analytics.native.ts
@@ -1,0 +1,1 @@
+export { analytics } from '../suite-native';
