+++ suite-common/analytics/src/events/shared/analytics.ts
@@ -1,0 +1,1 @@
+export { analytics } from '../suite';
