+++ suite-common/analytics/src/events/suite-native/constants.ts
