+++ suite-common/analytics/src/events/suite-native/definitions.ts
