+++ suite-common/analytics/src/events/suite/constants.ts
@@ -117,18 +117,5 @@ export enum EventType {
 
     T3T1DashboardPromo = 'promo/t3t1-dashboard',
 
-    WalletConnectInit = 'wallet-connect/init',
-    WalletConnectPaired = 'wallet-connect/paired',
-    WalletConnectProposal = 'wallet-connect/proposal',
-    WalletConnectProposalApproved = 'wallet-connect/proposal-approved',
-    WalletConnectProposalRejected = 'wallet-connect/proposal-rejected',
-    WalletConnectSessionRequest = 'wallet-connect/session-request',
-    WalletConnectError = 'wallet-connect/error',
-
-    ConnectPopupInit = 'connect-popup/init',
-    ConnectPopupPermissions = 'connect-popup/permissions',
-    ConnectPopupCall = 'connect-popup/call',
-    ConnectPopupError = 'connect-popup/error',
-
     AutostartModal = 'autostart-modal',
 }
