+++ suite-common/analytics/src/redux/analyticsActions.ts
