+++ suite-common/analytics/src/redux/analyticsReducer.ts
