+++ suite-common/analytics/tsconfig.json
@@ -1,5 +1,15 @@
 {
     "extends": "../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
-    "references": [{ "path": "../redux-utils" }]
+    "references": [
+        { "path": "../redux-utils" },
+        { "path": "../suite-config" },
+        { "path": "../suite-constants" },
+        { "path": "../trading" },
+        { "path": "../wallet-config" },
+        { "path": "../wallet-types" },
+        { "path": "../../packages/analytics" },
+        { "path": "../../packages/connect" },
+        { "path": "../../packages/env-utils" }
+    ]
 }
