+++ suite-native/analytics/README.md
@@ -1,5 +1,5 @@
-# Suite Analytics
+# Suite Ananlytics - Native
 
-This package is intended to be used only by Suite in web (or in desktop app renderer) environment.
+This package is intended to be used only by Suite in native (mobile) environment.
 
 It does not contain events by itself, please refer to [`@suite-common/analytics`](../../suite-common/analytics/README.md) for more information.
