+++ suite-native/analytics/package.json
@@ -12,14 +12,8 @@
     "dependencies": {
         "@suite-common/analytics": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
-        "@suite-common/suite-config": "workspace:*",
-        "@suite-common/suite-constants": "workspace:*",
-        "@suite-common/trading": "workspace:*",
-        "@suite-common/wallet-config": "workspace:*",
-        "@suite-common/wallet-types": "workspace:*",
         "@suite-native/config": "workspace:*",
         "@trezor/analytics": "workspace:*",
-        "@trezor/connect": "workspace:*",
         "@trezor/env-utils": "workspace:*"
     }
 }
