--- suite-native/analytics/src/analytics.ts
@@ -1,20 +1,0 @@
-import { isDebugEnv } from '@suite-native/config';
-import { Analytics, Event } from '@trezor/analytics';
-import { getSuiteVersion } from '@trezor/env-utils';
-
-import { SuiteNativeAnalyticsEvent } from './events';
-
-export const analytics = new Analytics<SuiteNativeAnalyticsEvent>({
-    version: getSuiteVersion(),
-    app: 'suite',
-});
-
-if (isDebugEnv()) {
-    // Do not send analytics in development
-    analytics.report = (event: Event) => {
-        if (process.env.EXPO_PUBLIC_IS_ANALYTICS_LOGGER_ENABLED === 'true') {
-            // eslint-disable-next-line no-console
-            console.log(`Analytics report '${event.type}':`, event);
-        }
-    };
-}
