+++ suite-native/analytics/tsconfig.json
@@ -8,22 +8,8 @@
         {
             "path": "../../suite-common/redux-utils"
         },
-        {
-            "path": "../../suite-common/suite-config"
-        },
-        {
-            "path": "../../suite-common/suite-constants"
-        },
-        { "path": "../../suite-common/trading" },
-        {
-            "path": "../../suite-common/wallet-config"
-        },
-        {
-            "path": "../../suite-common/wallet-types"
-        },
         { "path": "../config" },
         { "path": "../../packages/analytics" },
-        { "path": "../../packages/connect" },
         { "path": "../../packages/env-utils" }
     ]
 }
